requirements.txt
asyncpg 
redis>=4.2.0
python-dotenv 
pytest 
pytest-asyncio
sqlalchemy